backend/ai/requirements.txt
flask==3.0.2
textblob==0.18.0
pillow==10.2.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.4
Werkzeug==3.0.1
prophet
